Add files via upload
The resetpassword.py script generates a random password, and emails it to the email address provided by the user. It first looks up whether there is a user in UserCol that matches this email address, and if there is not, it prints a message stating as such and then exits. If there is such a user, the script will update the password for that user's entry, then send an automatically-generated email to the email address for that user, telling them what their new (randomized) password is.

The hash_password() function uses bcrypt, but it might not function the same way that the hashing function used in Danny's code does; if that proves to be the case, feel free to change my hash_password() function, and please let me know what to change in order to make it fit the existing code.